Update the Spring Cloud Function sample to Spring Boot 3.2.0 and sync it with the documentation
