+++ samples/spring-cloud-example/src/main/java/com/example/hello/HelloHandler.java
@@ -1,10 +1,10 @@
-package example.hello;
+package com.example.hello;
 
 import com.microsoft.azure.functions.*;
 import com.microsoft.azure.functions.annotation.AuthorizationLevel;
 import com.microsoft.azure.functions.annotation.FunctionName;
 import com.microsoft.azure.functions.annotation.HttpTrigger;
-import example.hello.model.*;
+import com.example.model.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -18,14 +18,12 @@ public class HelloHandler {
 
     @FunctionName("hello")
     public HttpResponseMessage execute(
-            @HttpTrigger(name = "request", methods = {HttpMethod.GET, HttpMethod.POST}, authLevel = AuthorizationLevel.ANONYMOUS) HttpRequestMessage<Optional<User>> request,
-            ExecutionContext context) {
+            @HttpTrigger(name = "request", methods = {HttpMethod.GET, HttpMethod.POST}, authLevel = AuthorizationLevel.ANONYMOUS) HttpRequestMessage<Optional<User>> request, ExecutionContext context) {
         User user = request.getBody()
                 .filter(u -> u.getName() != null)
                 .orElseGet(() -> new User(request.getQueryParameters().getOrDefault("name", "world")));
         context.getLogger().info("Greeting user name: " + user.getName());
-        return request
-                .createResponseBuilder(HttpStatus.OK)
+        return request.createResponseBuilder(HttpStatus.OK)
                 .body(hello.apply(user))
                 .header("Content-Type", "application/json")
                 .build();
