+++ samples/spring-cloud-example/src/main/java/com/example/model/Greeting.java
@@ -1,4 +1,4 @@
-package example.hello.model;
+package com.example.model;
 
 public class Greeting {
