+++ samples/spring-cloud-example/src/main/java/com/example/model/User.java
@@ -1,4 +1,4 @@
-package example.hello.model;
+package com.example.model;
 
 public class User {
