--- samples/spring-cloud-example/src/main/java/example/Application.java
@@ -1,12 +1,0 @@
-package example;
-
-import example.uppercase.Config;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-    public static void main(String[] args) throws Exception {
-        SpringApplication.run(Config.class, args);
-    }
-}
