--- samples/spring-cloud-example/src/main/java/example/uppercase/EchoHandler.java
@@ -1,31 +1,0 @@
-package example.uppercase;
-
-import com.microsoft.azure.functions.ExecutionContext;
-import com.microsoft.azure.functions.HttpMethod;
-import com.microsoft.azure.functions.HttpRequestMessage;
-import com.microsoft.azure.functions.annotation.AuthorizationLevel;
-import com.microsoft.azure.functions.annotation.FunctionName;
-import com.microsoft.azure.functions.annotation.HttpTrigger;
-
-import java.util.Optional;
-import java.util.function.Function;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cloud.function.adapter.azure.FunctionInvoker;
-import org.springframework.messaging.Message;
-import org.springframework.messaging.support.MessageBuilder;
-import org.springframework.stereotype.Component;
-
-@Component
-public class EchoHandler {
-
-	@Autowired
-	private Function<String, String> echo;
-
-	@FunctionName("echo")
-	public String execute(@HttpTrigger(name = "req", methods = {HttpMethod.GET,
-			HttpMethod.POST}, authLevel = AuthorizationLevel.ANONYMOUS) HttpRequestMessage<Optional<String>> request,
-		ExecutionContext context) {
-		return echo.apply(request.getBody().get());
-	}
-}
