+++ samples/spring-cloud-example/src/main/resources/host.json
@@ -2,6 +2,7 @@
 	"version": "2.0",
 	"extensionBundle": {
 		"id": "Microsoft.Azure.Functions.ExtensionBundle",
-		"version": "[4.*, 5.0.0)"
-	}
+		"version": "[4.*, 5.2.0)"
+	},
+	"functionTimeout": "00:10:00"
 }
