+++ samples/spring-cloud-example/src/test/java/com/example/HelloTest.java
@@ -1,0 +1,17 @@
+package com.example;
+
+import com.example.hello.Hello;
+import com.example.model.Greeting;
+import com.example.model.User;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class HelloTest {
+
+    @Test
+    public void test() {
+        Greeting result = new Hello().apply(new User("foo"));
+        assertThat(result.getMessage()).isEqualTo("Hello, foo!\n");
+    }
+}
